requirements.txt
astro-sdk-python
dbt-snowflake==1.3.0b1  ## current GA version is 1.2 ; this is pinned so that astro-sdk-python package can also be installed in the same Docker
apache-airflow-providers-snowflake
openlineage-dbt
apache-airflow-providers-sendgrid
apache-airflow-providers-common-sql


## for using Vault
apache-airflow-providers-hashicorp

## for wasb
apache-airflow-providers-microsoft-azure==4.3.0

## only for pip-compile/venv testing
# pip-tools
# virtualenv

astronomer-aeroscope